# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EFFF Club Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webinars:
    get:
      operationId: listWebinars
      tags: [webinars]
      summary: List webinars and seminars
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, past]
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webinar"
    post:
      operationId: createWebinar
      tags: [webinars]
      summary: Create a webinar (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebinarInput"
      responses:
        "201":
          description: Webinar created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webinar"
...
  /webinars/{id}/register:
    post:
      operationId: registerForWebinar
      tags: [webinars]
      summary: Register for a webinar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # FINANCE
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Webinar:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [webinar, seminar]
        speaker:
          type: string
        date:
          type: string
        duration:
          type: integer
        price:
          type: number
        registeredCount:
          type: integer
        maxCapacity:
          type: integer
        isRegistered:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - type
        - speaker
        - date
        - duration
        - price
        - registeredCount
        - maxCapacity
        - isRegistered
        - createdAt
...
    WebinarInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [webinar, seminar]
        speaker:
          type: string
        date:
          type: string
        duration:
          type: integer
        price:
          type: number
        maxCapacity:
          type: integer
        location:
          type: string
      required:
        - title
        - description
        - type
        - speaker
        - date
        - duration
        - price
        - maxCapacity
...